src/student-requirements-pip.txt
websocket-client~=1.8.0
rel~=0.4.9.19